Remove cargo frozen flag
Even if the Cargo.lock file is unchanged when running builds locally and
all Cargo.toml files are copied correctly in the Docker build, I still
see error messages like this every now and then:

    error: failed to get `chrono` as a dependency of package `ghss_github v0.1.0 (/src/crates/ghss_github)`

    Caused by:
      failed to load source for dependency `chrono`

    Caused by:
      Unable to update registry `https://github.com/rust-lang/crates.io-index`

    Caused by:
      attempting to make an HTTP request, but --frozen was specified

I'm unsure what's causing this. In any case, --frozen doesn't seem to be
important enough to deal with this right now.